+++ assets/controllers/sync_controller.js
@@ -1,10 +1,6 @@
 import { Controller } from "@hotwired/stimulus";
 
-/**
- * Project sync controller.
- *
- * Sync a project
- */
+/** Sync controller. */
 export default class extends Controller {
     static targets = ["text", "button"];