+++ src/Command/QueueSyncCommand.php
@@ -6,14 +6,12 @@ use App\Entity\SynchronizationJob;
 use App\Enum\SynchronizationStatusEnum;
 use App\Message\SynchronizeMessage;
 use App\Repository\SynchronizationJobRepository;
-use Psr\Log\LoggerInterface;
 use Symfony\Component\Console\Attribute\AsCommand;
 use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Output\OutputInterface;
 use Symfony\Component\Messenger\MessageBusInterface;
 use Symfony\Component\Scheduler\Attribute\AsCronTask;
-use Symfony\Contracts\HttpClient\HttpClientInterface;
 
 #[AsCommand(
     name: 'app:queue-sync',
@@ -25,9 +23,6 @@ class QueueSyncCommand extends Command
     public function __construct(
         private readonly SynchronizationJobRepository $synchronizationJobRepository,
         private readonly MessageBusInterface $bus,
-        private readonly HttpClientInterface $client,
-        private readonly LoggerInterface $logger,
-        private readonly string $monitoringUrl,
     ) {
         parent::__construct($this->getName());
     }
@@ -38,15 +33,6 @@ class QueueSyncCommand extends Command
 
     protected function execute(InputInterface $input, OutputInterface $output): int
     {
-        // Call monitoring url if defined.
-        if ('' !== $this->monitoringUrl) {
-            try {
-                $this->client->request('GET', $this->monitoringUrl);
-            } catch (\Throwable $e) {
-                $this->logger->error('Error calling monitoringUrl: '.$e->getMessage());
-            }
-        }
-
         $job = new SynchronizationJob();
         $job->setStatus(SynchronizationStatusEnum::NOT_STARTED);
         $this->synchronizationJobRepository->save($job, true);