+++ src/Controller/HomeController.php
@@ -2,6 +2,8 @@
 
 namespace App\Controller;
 
+use App\Entity\User;
+use App\Service\DashboardService;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Attribute\Route;
@@ -14,8 +16,16 @@ class HomeController extends AbstractController
     }
 
     #[Route('/', name: 'index')]
-    public function index(): Response
+    public function index(DashboardService $dashboardService): Response
     {
-        return $this->render('home/index.html.twig');
+        /** @var User $user */
+        $user = $this->getUser();
+
+        $dashboardData = $dashboardService->getUserDashboard($user);
+
+        return $this->render('home/index.html.twig', [
+            'userName' => $user->getName(),
+            'dashboardData' => $dashboardData,
+        ]);
     }
 }