+++ src/Controller/InvoicingRateReportController.php
@@ -37,7 +37,7 @@ class InvoicingRateReportController extends AbstractController
             'action' => $this->generateUrl('app_invoicing_rate_report'),
             'method' => 'GET',
             'attr' => [
-                'id' => 'sprint_report',
+                'id' => 'report',
             ],
             'years' => [
                 (new \DateTime())->modify('-1 year')->format('Y'),