+++ src/Form/BillableUnbilledHoursReportType.php
@@ -46,9 +46,6 @@ class BillableUnbilledHoursReportType extends AbstractType
     {
         $resolver->setDefaults([
             'data_class' => BillableUnbilledHoursReportFormData::class,
-            'attr' => [
-                'data-sprint-report-target' => 'form',
-            ],
             'years' => null,
         ]);
     }