+++ src/Form/PlanningType.php
@@ -2,7 +2,9 @@
 
 namespace App\Form;
 
+use App\Entity\WorkerGroup;
 use App\Model\Planning\PlanningFormData;
+use Symfony\Bridge\Doctrine\Form\Type\EntityType;
 use Symfony\Component\Form\AbstractType;
 use Symfony\Component\Form\Extension\Core\Type\ChoiceType;
 use Symfony\Component\Form\FormBuilderInterface;
@@ -22,9 +24,19 @@ class PlanningType extends AbstractType
                 'help_attr' => ['class' => 'form-help'],
                 'row_attr' => ['class' => 'form-row'],
                 'required' => false,
-                'data' => $currentYear,
+                'empty_data' => $currentYear,
                 'choices' => [$currentYear => $currentYear, $nextYear => $nextYear],
                 'placeholder' => null,
+            ])
+            ->add('group', EntityType::class, [
+                'class' => WorkerGroup::class,
+                'label' => 'planning.group',
+                'label_attr' => ['class' => 'label'],
+                'attr' => ['class' => 'form-element ', 'onchange' => 'this.form.submit()'],
+                'help_attr' => ['class' => 'form-help'],
+                'row_attr' => ['class' => 'form-row'],
+                'required' => false,
+                'placeholder' => 'planning.select_group',
             ])
         ;
     }