+++ src/Form/WorkerType.php
@@ -3,6 +3,8 @@
 namespace App\Form;
 
 use App\Entity\Worker;
+use App\Entity\WorkerGroup;
+use Symfony\Bridge\Doctrine\Form\Type\EntityType;
 use Symfony\Component\Form\AbstractType;
 use Symfony\Component\Form\Extension\Core\Type\ChoiceType;
 use Symfony\Component\Form\Extension\Core\Type\TextType;
@@ -51,6 +53,17 @@ class WorkerType extends AbstractType
                 ],
                 'multiple' => false,
                 'placeholder' => false,
+            ])
+            ->add('workerGroups', EntityType::class, [
+                'class' => WorkerGroup::class,
+                'multiple' => true,
+                'by_reference' => false,
+                'label' => 'worker.groups',
+                'label_attr' => ['class' => 'label'],
+                'attr' => ['class' => 'form-element', 'data-choices-target' => 'choices'],
+                'help_attr' => ['class' => 'form-help'],
+                'required' => false,
+                'row_attr' => ['class' => 'form-row'],
             ])
         ;
     }