+++ src/Model/Planning/Issue.php
@@ -7,10 +7,10 @@ class Issue
     public readonly string $key;
     public readonly string $displayName;
     public readonly ?float $remainingHours;
-    public readonly string $link;
+    public readonly ?string $link;
     public readonly string $sprintId;
 
-    public function __construct(string $key, string $displayName, ?float $remainingHours, string $link, string $stringId)
+    public function __construct(string $key, string $displayName, ?float $remainingHours, ?string $link, string $stringId)
     {
         $this->key = $key;
         $this->displayName = $displayName;